🍋👻 Checkpoint ./src/state/search.js:271885/11958 ./src/presenters/team-item.js:271885/44 ./src/presenters/pages/search-new.js:271885/261 ./src/presenters/pages/router.js:271885/4 ./src/presenters/includes/user-prefs.js:271885/40 ./src/presenters/pages/search.js:271885/219
